Add unit tests for moved helper methods
