+++ tests/Cms/Helpers/fixtures/HelpersTest/snippet-item.php
@@ -1,0 +1,1 @@
+Hello <?= $item->method() ?>
