+++ tests/Filesystem/fixtures/f/load/B/B.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace FTest;
+
+class B
+{
+}
