+++ tests/Filesystem/fixtures/f/load/a/a.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace FTest;
+
+class A
+{
+}
